#7 Create the host (mediator) and entity
This is the start of both the Host and the Entity. Both are not fully complete yet, but are basically functional in their current state. I also moved the window code into the host so that it wasn't floating around in the main. I will need to move the window code into it's own container as we can have multiple windows.