#22 Adding mouse window events to X11
This is working X11 mouse input events. There is also a bit of refactor for the windows mouse events here to make more common code between the windowing systems. I refactored the windows code so that which button was being clicked is being sent rather than just X1 and X2