#15 Setup yaegi Go interpreter
This is a functional implementation of the yaegi Go interpreter for interpreting Go code as a scripting language for the engine. This is built so that when the release build of the game/application is built, the go code is compiled directly into the executable rather than being interpreted. I've also added the code needed to execute the project source code at any time on a given host.