#37 Hook up GL calls to main go thread
Making the setup of a window broken up into 3 phases. (1) Create the window, (2) create the context in go, (3) start the main window event loop. This uses the shared memory to facilitate the stages between C and Go. It's not completely ideal, but it is functional and not too much of a change from the initial implementation. I also moved the glad and GL specific code into the gl package (folder) in the project. Extra helper methods for knowing the state and setting a fatal have been added to shared memory.