Shader load & triangle draw in OpenGL on Win32/X11
This creates the base concept of a shader for OpenGL. It also creates a shader cache to hold shaders that have been built for reuse. This also creates the stub for all GL renderering through a C wrapper layer. Lastly a triangle is crudely drawn on the screen to verify the shader load/compile/use.