#40 Create standard vertex layout for rendering
This creates a Vertex structure so and the associated code to load it up for rendering. The renderer is modified to have the create mesh function. Meshes are bare minimum right now as the focus was on the vertex layout. I also added Color, Color8, and Vec4i helper structs into the matrix library.