#49 Cleanup rendering, make instanced draw default
This changes the sample code in main to use instanced drawing and instanced data controlled through an instance group. This is not a complete finish of the implementation as we'll need to share shaders and only bind the shader once. This will likely be through a shader draw which then has all the instance groups beneath it. For now the instancing and writing of instance data into a texture for GL to use in the shader is functional.